Fix error; no longer throw errors about unknown targets.
The error string wasn't handling 'None' properly; fixed that.

Also, instead of skipping all targets with "/@" in them (i.e. xpaths to attributes), we try to find a target anyway, and if so, great, and if not, silently don't add it to the map; we'll fail later when trying to get a value for it.